RavenDB-3947: Activate compression from the client side. In the current subscriptions and bulk loading test scenario the amount of data transferred through the network shrink by approximately 70%.
