+++ Raven.Database/FileSystem/Storage/Voron/StorageActionsAccessor.cs
@@ -25,6 +25,7 @@ using RavenConstants = Raven.Abstractions.Data.Constants;
 
 using Raven.Abstractions.FileSystem;
 using Raven.Abstractions.Data;
+using Raven.Abstractions.Util;
 
 namespace Raven.Database.FileSystem.Storage.Voron
 {
