+++ Raven.Database/Raven.Database.csproj
@@ -1155,7 +1155,6 @@
     <Compile Include="Storage\Voron\StorageActions\StorageActionsBase.cs" />
     <Compile Include="Storage\Voron\StorageActions\TasksStorageActions.cs" />
     <Compile Include="Storage\Voron\TransactionalStorage.cs" />
-    <Compile Include="Storage\Voron\UndisposableStream.cs" />
     <Compile Include="Storage\TransactionContextData.cs" />
     <Compile Include="Tasks\RemoveFromIndexTask.cs" />
     <Compile Include="Tasks\DatabaseTask.cs" />
